feat(ui): merge Macro and Combo tabs into a single Macro&Combo tab
Combine the two runtime-editing tabs into one page with a shared layout:
Macros and Combos list cards stacked in the left column with exclusive
selection, a right column that switches between the macro editor, combo
editor, and per-domain global settings cards (opened via a gear button
next to each list's + button), and a unified Refresh/Discard/Save bar
that applies to both subsystems. Global-settings pending changes are now
tracked per field so green dots appear on both the gear button and the
corresponding row inside the settings card.
